+++ include/core/utils.h
@@ -14,9 +14,9 @@ There is NO WARRANTY, to the extent permitted by law.
 #ifdef _WIN32
 #   define WIN32_LEAN_AND_MEAN
 #   include "core/win32.h"
-#   include "exports/utils_exports.h"
+#   include "exports/core_exports.h"
 #else
-#   define UTILS_EXPORT
+#   define CORE_EXPORT
 #endif
 
 // these are orpg connonical exit and output codes
@@ -38,29 +38,29 @@ namespace ORPG {
          * This adhears to the GNU standard for version printing, and immediately terminates
          * the program with exit code EXIT_SUCCESS
          **/
-        void UTILS_EXPORT print_version_flag();
+        void CORE_EXPORT print_version_flag();
 
         /**
          * @desc prints the help info when -h or --help is an argument to the command.
          * This adhears to the GNU standard for help printing, and immediately terminates
          * the program with exit code EXIT_SUCCESS
          **/
-        void UTILS_EXPORT print_help_flag();
+        void CORE_EXPORT print_help_flag();
 
         /**
          * @desc prints the version info when version, ver, v, or V are called in the ORPG shell.
          * Because this is called from within our ORPG shell, the program will continue running.
          **/
-        void UTILS_EXPORT print_basic_version();
+        void CORE_EXPORT print_basic_version();
 
         /**
          * @desc prints the help info when help, h, or H are called in the ORPG shell.
          * Because this is called from within our ORPG shell, the program will continue running.
          **/
-        void UTILS_EXPORT print_basic_help();
+        void CORE_EXPORT print_basic_help();
 
-        extern bool UTILS_EXPORT VB_FLAG;
-        extern bool UTILS_EXPORT QUIET_FLAG;
+        extern bool CORE_EXPORT VB_FLAG;
+        extern bool CORE_EXPORT QUIET_FLAG;
     };
 };
 
@@ -70,40 +70,40 @@ namespace ORPG {
          * Loads the menu file of the given type to be used
          * for printing the UI
          **/
-        std::string UTILS_EXPORT get_display_screen(std::string file);
+        std::string CORE_EXPORT get_display_screen(std::string file);
 
         /**
          * Converts the given file to an exact string copy
          * used to create images and other printed files.
          **/
-        std::string UTILS_EXPORT file_to_string(std::string fileName);
+        std::string CORE_EXPORT file_to_string(std::string fileName);
 
         // Taken from https://stackoverflow.com/questions/6089231/getting-std-ifstream-to-handle-lf-cr-and-crlf
-        std::istream UTILS_EXPORT & safeGetline(std::istream& is, std::string& t);
+        std::istream CORE_EXPORT & safeGetline(std::istream& is, std::string& t);
 
         /**
          * Prints the text contents of the given file to stdout
          **/
-        bool UTILS_EXPORT print_file(std::string file);
+        bool CORE_EXPORT print_file(std::string file);
 
         /**
          * Example: str = "hi", len = "5", ch = '-'
          * Return: "---hi"
          **/
-        std::string UTILS_EXPORT leftpad(std::string str, int len = 0, char ch = ' ');
+        std::string CORE_EXPORT leftpad(std::string str, int len = 0, char ch = ' ');
 
         /**
          * Example: str = "hi", len = "5", ch = '-'
          * Return: "hi---"
          **/
-        std::string UTILS_EXPORT rightpad(std::string str, int len = 0, char ch = ' ');
+        std::string CORE_EXPORT rightpad(std::string str, int len = 0, char ch = ' ');
 
-        int UTILS_EXPORT randomInt(int min, int max);
+        int CORE_EXPORT randomInt(int min, int max);
 
-        bool UTILS_EXPORT randomBool();
+        bool CORE_EXPORT randomBool();
 
         /* Compute the greatest common divisor of a and b. */
-        int UTILS_EXPORT gcd(int a, int b);
+        int CORE_EXPORT gcd(int a, int b);
     }
 }