+++ src/character/character.cpp
@@ -231,7 +231,7 @@ namespace ORPG {
          * by one.
          * 
          * TODO(incomingstick): Make this generic and put it somewhere in the Utils
-         * library
+         * namespace
          *
          * @param vector<uint8> *arr - the pointer of the vectory array to operate on
          * @return auto - the extracted value from the randomly selected element
@@ -495,8 +495,8 @@ namespace ORPG {
 
         /**
          * @desc This function prompts the user via stdout for a name, and reading
-         * from stdin the input. We use the safeGetline funtion via the Utils
-         * library to ensure integrity across platforms.
+         * from stdin the input. We use the safeGetline funtion via the ORPG::Utils
+         * namespace to ensure integrity across platforms.
          * 
          * @return string - the user input string to be used as a name
          **/